% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricv.R
\name{ricv}
\alias{ricv}
\title{Create an interactive image comparator}
\usage{
ricv(img1, img2, options = NULL, css = NULL)
}
\arguments{
\item{img1}{A path to an image file. This image will be placed to the left (i.e. 'before' image).}

\item{img2}{A path to an image file. This image will be placed to the right (i.e. 'after' image).}

\item{options}{A list of image-compare-viewer options. See documentation for all options.}

\item{css}{A list of 3 optional attributes ('both', 'before' and 'after'), each containing CSS style attributes targeting the label corresponding to the used attribute name.}
}
\description{
This function creates an interactive image comparator. See website of the
underlying project for examples: \url{https://image-compare-viewer.netlify.app/}
}
